[svn-r5326] Purpose:  code clean-up Description: Many warnings are generated on windows due to seemingly large-size data type converting to small-size data type, 1.5 branch has been cleaned up, make 1.4 catch up with 1.5. Solution: If meeting with the break-up of the current tests on other platforms, will resume the previous one Platforms tested: windows 2000, linux 2.2.18

+++ src/H5Smpio.c
@@ -105,7 +105,7 @@ H5S_mpio_all_type( const H5S_t *space, const size_t elmt_size,
 
     /* fill in the return values */
     *new_type = MPI_BYTE;
-    *count = total_bytes;
+    *count = (size_t)total_bytes;
     *is_derived_type = 0;
 
 #ifdef H5Smpi_DEBUG
